api/requirements.txt
# framework
fastapi[standard]
# configs
pyyaml
# llm providers
openai
# requests
jsons
requests
python-multipart
# parsing
torch
torchvision
marker-pdf==1.3.5
Pillow
docling
# unstructured
langchain-unstructured
# database
neo4j

gunicorn

uvicorn # remove later
gunicorn
ipykernel
termcolor 
python-dotenv